adding routers
