+++ package.json
@@ -14,6 +14,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.0",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.2"
   },
