+++ src/App.css
@@ -2,7 +2,7 @@
 
 .App {
   min-height: 100vh;
-  background-image: linear-gradient(to right, #FFffff 0%, #B4CEFF 100%);
+  background-image: linear-gradient(to right, #ffffff 0%, #b4ceff 100%);
 }
 .search-form {
   min-height: 10vh;
@@ -30,7 +30,7 @@
   transition: 0.5s;
 }
 .search-bar:focus {
-  outline: 5px solid #4B3100;
+  outline: 5px solid #4b3100;
 }
 .search-form:hover {
   transform: scale(1.1);
@@ -39,7 +39,7 @@
   background: rgb(255, 238, 170);
 }
 .search-btn {
-  background: #4B3100;
+  background: #4b3100;
   border: none;
   color: white;
   padding: 10px 20px;
@@ -87,21 +87,22 @@
   border: none;
   border-radius: 10px;
   transition: 0.3s;
-  background-color: #B193AF;
-  
+  background-color: #eecdec;
+  color: black;
+  font-weight: 500;
 }
 .see-recipe-btn:hover {
   transform: scale(1.1);
-  background-color: #B193AF;
+  background-color: #b193af;
 }
 
 .loadMoreButton {
   border-radius: 10px;
-  height:60px;
+  height: 60px;
   width: 200px;
-  background-color: #4B3100;
+  background-color: #4b3100;
   color: white;
-  border-color: #4B3100;
+  border-color: #4b3100;
   font-size: medium;
 }
 
@@ -110,13 +111,11 @@ body {
   font-family: "Poppins", sans-serif;
 }
 .logo {
-    display: block;
-    margin: 0 auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 20px;
-    margin-top: 50px;
+  display: block;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 20px;
+  margin-top: 50px;
 }
-
-
