+++ src/Pages/Contact.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Contact() {
+  return <div>This Contact page</div>;
+}
